refactor: remove containers
